CODE RUB: Renaming Retrieve to RetrieveAll and RetrieveById and splitting test scenarios
