+++ OtripleS.Web.Api.Tests.Unit/Services/Foundations/Teachers/TeacherServiceTests.Exceptions.RetrieveById.cs
@@ -17,74 +17,6 @@ namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.Teachers
     public partial class TeacherServiceTests
     {
         [Fact]
-        public void ShouldThrowDependencyExceptionOnRetrieveAllWhenSqlExceptionOccursAndLogIt()
-        {
-            // given
-            SqlException sqlException = GetSqlException();
-
-            var expectedTeacherDependencyException =
-                new TeacherDependencyException(sqlException);
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.SelectAllTeachers())
-                    .Throws(sqlException);
-
-            // when . then
-            Assert.Throws<TeacherDependencyException>(() =>
-                this.teacherService.RetrieveAllTeachers());
-
-            this.loggingBrokerMock.Verify(broker =>
-                broker.LogCritical(It.Is(SameExceptionAs(expectedTeacherDependencyException))),
-                    Times.Once);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectAllTeachers(),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTime(),
-                    Times.Never);
-
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-            this.storageBrokerMock.VerifyNoOtherCalls();
-        }
-
-        [Fact]
-        public void ShouldThrowServiceExceptionOnRetrieveAllWhenExceptionOccursAndLogIt()
-        {
-            // given
-            var exception = new Exception();
-
-            var expectedTeacherServiceException =
-                new TeacherServiceException(exception);
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.SelectAllTeachers())
-                    .Throws(exception);
-
-            // when . then
-            Assert.Throws<TeacherServiceException>(() =>
-                this.teacherService.RetrieveAllTeachers());
-
-            this.loggingBrokerMock.Verify(broker =>
-                broker.LogError(It.Is(SameExceptionAs(expectedTeacherServiceException))),
-                    Times.Once);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectAllTeachers(),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTime(),
-                    Times.Never);
-
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-            this.storageBrokerMock.VerifyNoOtherCalls();
-        }
-
-        [Fact]
         public async Task ShouldThrowDependencyExceptionOnRetrieveWhenSqlExceptionOccursAndLogItAsync()
         {
             // given
